Added SonarQube and refactor of if statements

+++ src/main/java/com/wt/service/SetService.java
@@ -21,6 +21,8 @@ public class SetService {
     @Inject
     SecurityIdentity securityIdentity;
 
+    private static final String INVALIDUSER = "Error: Invalid user";
+
     public List<Set> allSets() {
         return setRepository.listAll();
     }
@@ -31,30 +33,27 @@ public class SetService {
     }
 
     public String addSet(UUID exerciseId, String userEmail) {
-        if (!userService.checkIfUserExists(userEmail)) {
-            return "User does not exist";
-        }
-        if (!securityIdentity.getPrincipal().getName().equals(userEmail)) {
-            return "User not the same as token user";
+        if (!userService.checkIfUserExists(userEmail) || !securityIdentity.getPrincipal().getName().equals(userEmail)) {
+            throw new IllegalArgumentException(INVALIDUSER);
         }
         if (exerciseId == null) {
-            return "ExerciseId is not defined";
+            throw new IllegalArgumentException("exerciseId is not defined");
         }
-
-        Set set = new Set(exerciseId, 0, 0, 0, 0);
-        setRepository.persist(set);
-        return "Set created";
+        try {
+            Set set = new Set(exerciseId, 0, 0, 0, 0);
+            setRepository.persist(set);
+            return "Set created";
+        } catch (Exception error) {
+            return error.getMessage();
+        }
     }
 
     public String updateSet(UUID setId, int sets, int reps, double weight, int rest, String userEmail) {
-        if (!userService.checkIfUserExists(userEmail)) {
-            return "User does not exist";
-        }
-        if (!securityIdentity.getPrincipal().getName().equals(userEmail)) {
-            return "User not the same as token user";
+        if (!userService.checkIfUserExists(userEmail) || !securityIdentity.getPrincipal().getName().equals(userEmail)) {
+            throw new IllegalArgumentException(INVALIDUSER);
         }
         if (setId == null) {
-            return "ExerciseId is not defined";
+            throw new IllegalArgumentException("setId is not defined");
         }
 
         try {
@@ -66,23 +65,20 @@ public class SetService {
                             .and("id", setId));
             return "Set updated";
         } catch (Exception error) {
-            return String.format("{\"message\": \"%s\"}", error.getMessage());
+            return error.getMessage();
         }
     }
 
     public String deleteSet(UUID id, String userEmail) {
-        if (!userService.checkIfUserExists(userEmail)) {
-            return "User does not exist";
-        }
-        if (!securityIdentity.getPrincipal().getName().equals(userEmail)) {
-            return "User does not equal token";
+        if (!userService.checkIfUserExists(userEmail) || !securityIdentity.getPrincipal().getName().equals(userEmail)) {
+            throw new IllegalArgumentException(INVALIDUSER);
         }
 
         try {
             setRepository.delete("id", id);
             return "Set is deleted";
         } catch (Exception error) {
-            return "Something went wrong";
+            return error.getMessage();
         }
     }
 }
